producer_news/requirements.txt
pandas>=1.5.2
beautifulsoup4>=4.11.1
requests>=2.28.1
numpy>=1.24.0
pytz>=2022.7
kafka-python>=2.0.2
